requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms==2.3.3
coverage
geopy
geocoder
# ecdsa
# paramiko
# pycrypto
